Cleanup store initialization (#8156)
* assume enableThunk is never false

* drop unused clientOptions

* inline createMiddleware

* collapse prod/dev configureStore

* typescript magic

* redux-devtools-extension/developmentOnly

* fix linting
