+++ package.json
@@ -133,6 +133,7 @@
     "@types/redux-mock-store": "1.0.2",
     "@types/shallow-equals": "1.0.0",
     "@types/tinycolor2": "1.4.2",
+    "@types/webpack-env": "1.16.0",
     "@typescript-eslint/eslint-plugin": "4.22.0",
     "@typescript-eslint/parser": "4.22.0",
     "babel-eslint": "10.1.0",
