+++ packages/mattermost-redux/src/store/index.ts
@@ -1,6 +1,5 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-/* eslint-disable global-require, no-process-env */
-const config = process.env.NODE_ENV === 'production' ? require('./configureStore.prod').default : require('./configureStore.dev').default;
+const config = require('./configureStore').default;
 export default config;
